requirements.txt
absl-py==0.12.0
alabaster==0.7.12
appnope==0.1.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==1.5.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
entrypoints==0.3
Fiona==1.8.19
flatbuffers==1.12
gast==0.3.3
geopandas==0.9.0
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
imgviz==1.2.6
importlib-metadata==3.8.1
ipykernel==5.5.0
ipyparallel==6.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
Keras==2.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme==3.16.5
Markdown==3.3.4
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.2.2
mistune==0.8.4
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nose==1.3.7
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
osgeo==0.0.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyproj==3.0.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.17.2
scipy==1.5.4
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
snowballstemmer==2.1.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==1.8.0
tensorboard-plugin-wit==1.8.0
tensorflow==1.3.0
tensorflow-estimator==2.4.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
text-unidecode==1.3
tifffile==2020.9.3
tinytools==1.1.0
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
tzwhere==3.0.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.4.1
